realtor-dashboard-backend/ai-voice-assistant/requirements.txt
# Core Flask dependencies
Flask==3.1.0
flask-sock==0.7.0
Werkzeug==3.1.3

# Database
psycopg2-binary==2.9.10

# Authentication and security
PyJWT==2.10.1
python-dotenv==1.1.0

# Twilio for call handling
twilio==9.5.2

# NLP and AI
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.4
spacy==3.8.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

# Google Cloud services for speech
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1

# Elevenlabs for voice synthesis
elevenlabs==2.1.0
cartesia==2.0.3


# Ngrok for tunneling
pyngrok==7.2.3

# Websocket handling
simple-websocket==1.1.0
websockets==15.0.1
wsproto==1.2.0
webrtcvad==2.0.10

# Other essential utilities
typing_extensions==4.13.2

google-cloud-aiplatform==1.71.1
vertexai==1.71.1
sentencepiece==0.2.0
tokenizers==0.15.2